# gobackend/docs/swagger.yaml
basePath: /api
definitions:
...
  utils.ErrorInfo:
    properties:
      code:
        description: A machine-readable error code
        type: string
      details:
        additionalProperties:
          type: string
        description: Additional details about the error (e.g., validation errors)
        type: object
      message:
        description: A human-readable error message
        type: string
    type: object
  utils.MetaInfo:
    properties:
      page:
        description: The current page number
        type: integer
      page_size:
        description: The number of items per page
        type: integer
      total_items:
        description: The total number of items
        type: integer
      total_pages:
        description: The total number of pages
        type: integer
    type: object
  utils.Response:
    properties:
      data:
        description: The response data (omitted for error responses)
      error:
        allOf:
        - $ref: '#/definitions/utils.ErrorInfo'
        description: Error information (omitted for successful responses)
      meta:
        allOf:
        - $ref: '#/definitions/utils.MetaInfo'
        description: Metadata such as pagination information
      success:
        description: Whether the request was successful
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hidemeai.com
    name: API Support
    url: http://www.hidemeai.com/support
  description: Main function that starts the HideMe API Server
  license:
    name: Proprietary
    url: http://www.hidemeai.com/license
  termsOfService: http://www.hidemeai.com/terms/
  title: HideMe API
  version: 1.0.0
paths:
...
  /db/{table}:
    get:
      description: Returns all records from a table with optional filtering
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Records per page (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: Filter parameters (column names and values)
        in: query
        name: filter
        type: object
      produces:
      - application/json
      responses:
        "200":
          description: Data retrieved successfully with pagination
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid table name or parameters
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get table data
      tags:
      - Database
    post:
      consumes:
      - application/json
      description: Creates a new record in a table
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      - description: Record data
        in: body
        name: record
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Record created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid table name or request body
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Create record
      tags:
      - Database
  /db/{table}/{id}:
    delete:
      description: Deletes a record from a table
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Record deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Invalid table name or ID
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Delete record
      tags:
      - Database
    get:
      description: Returns a single record from a table by its ID
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Record retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid table name or ID
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get record by ID
      tags:
      - Database
    put:
      consumes:
      - application/json
      description: Updates an existing record in a table
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      - description: Record ID
        in: path
        name: id
        required: true
        type: string
      - description: Record data to update
        in: body
        name: record
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Record updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "400":
          description: Invalid table name, ID, or request body
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "404":
          description: Record not found
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Update record
      tags:
      - Database
  /db/{table}/schema:
    get:
      description: Returns the schema for a table
      parameters:
      - description: Table name
        in: path
        name: table
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Schema retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    additionalProperties: true
                    type: object
                  type: array
              type: object
        "400":
          description: Invalid table name
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "403":
          description: Table access not allowed
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server error
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                error:
                  type: string
              type: object
      security:
      - BearerAuth: []
      summary: Get table schema
      tags:
      - Database
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"